backend/myproject/urls.py
from django.contrib import admin
from django.urls import include, path
from myapp import views
from rest_framework.routers import DefaultRouter

router = DefaultRouter()
router.register(r'items', views.ItemViewSet, basename='item')

urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include(router.urls)),
    path('', include('django_prometheus.urls')),
]
