backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-prometheus==2.3.1
djangorestframework==3.16.0
prometheus_client==0.21.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
